+++ _content/projects/2020-mcgill-computer-taskforce.md
@@ -18,4 +18,4 @@ gallerySrcs:
   ]
 ---
 
-Led a team of developers to deliver the main website, coordinated the elections of 30 student groups during the pandemic, and instantiated a web based solution for the note takers club impacting all students in the faculty of Science & Arts and Science at McGill. To this day, all of these initiatives are still being extended.
+Led a team of developers to deliver the main website, coordinated the elections of 30 student groups during the pandemic, and instantiated a web based solution for the note takers club impacting all students in the faculty of Science & Arts and Science at McGill.
