+++ _content/projects/2023-garlic-festival.md
@@ -20,4 +20,4 @@ gallerySrcs:
   ]
 ---
 
-Coordinated the largest garlic festival in Quebec on a 3 month timeline with limited budget. In my time as lead I grew a 30 person team, negotiated partnerships with local businesses, and expanded our vendor roaster beyond what ever previously experienced in the last 15 years. Due to the festival's success, the event was covered by all major English radio stations in the Montreal area including CBC and CHOM and attracted ~7000 garlic enthusiasts.
+Coordinated the largest garlic festival in Quebec on a 3 month timeline with limited budget for ~7000 attendees. In my time as lead, I managed all aspects of event logistics, including vendor selection, sponsorship negotiation, media partnerships, budgeting, and on-site coordination.
