+++ src/components/layout/carousel/carousel.module.scss
@@ -1,0 +1,3 @@
+.carouselContainer {
+  display: flex;
+}
